+++ src/frontend/apps/impress/src/features/docs/doc-editor/hook/index.ts
@@ -1,0 +1,2 @@
+export * from './useSaveDoc';
+export * from './useUploadFile';
