+++ src/frontend/apps/impress/src/features/docs/doc-editor/hook/useUploadFile.tsx
@@ -1,0 +1,35 @@
+import { useCallback } from 'react';
+
+import { useMediaUrl } from '@/core/config';
+
+import { useCreateDocAttachment } from '../api';
+
+export const useUploadFile = (docId: string) => {
+  const mediaUrl = useMediaUrl();
+  const {
+    mutateAsync: createDocAttachment,
+    isError: isErrorAttachment,
+    error: errorAttachment,
+  } = useCreateDocAttachment();
+
+  const uploadFile = useCallback(
+    async (file: File) => {
+      const body = new FormData();
+      body.append('file', file);
+
+      const ret = await createDocAttachment({
+        docId,
+        body,
+      });
+
+      return `${mediaUrl}${ret.file}`;
+    },
+    [createDocAttachment, docId, mediaUrl],
+  );
+
+  return {
+    uploadFile,
+    isErrorAttachment,
+    errorAttachment,
+  };
+};
